Note that sqlite3 var is freed in finalization section
sqlite3 variable is freed automatically in finalization section of the unit,
if you use sqlite3.Free in the cleanup part (e.g. executed on closing your application)
you might get an AV because the object is already freed by the finalization code.
If you only use FreeAndNil to free it, you'll be safe in all cases as it's independent of 
the ordering of unloading the units and therefore execution of finalization code.