Update SynMongoDB.pas
Hi Arnaud,
with these changes (I've pasted all manually over Web, hope there are not too much mistakes) alle Connections of a Client Cluster could be used. The rpNearest is a new constant for Mongo 4.4 and I wanted to intruduce it for further using. Atm it should be a load balancing of using all Connections evenly. This could be usefull when many client are connecting to a Atlas Cluster in case of readings.
If you have questions, Email me.

Daniel